Fix serverinfo representation in nextcloud OCC command tool: add info command
